[doc checker] Fix convert-http-to-path bug
With fuchsia.git, links outside of //docs would always report an error:
if given as a relative path, an error would report that since it was
not in //docs it must be given as URL; if given as a URL, an error would
report that since the path lies within fuchsia.git, it must be given by
a relative path.

This change makes it so that http-to-path is only reuqested when the URL
that points within //docs.

Change-Id: I1bfbc74cf6adf2eb2c03bab571bb8b802fd65948Change-Id:
I62477d74c7826452586d9c7ec0de379d8bccc119

Change-Id: I528a8c8af51aefd4d43cfe5a2ebca1e22befd7af