[topaz/sdk] Implement VmoFile in Dart VFS library.
This change implements VmoFile as a specialization of PseudoFile in the
SDK. The current implementation is read-only and supports only two VMO
sharing modes (no sharing and duplicate, clone is not yet supported).

This change also makes the implementation of describe a method of the
PseudoFile so it can be overridden.

TEST=fuchsia_vfs_package_unittests

Change-Id: I0612d9619932c6043ae03c6b12c8890dded27fad