[dart] Expose svc handle to dart runtime
This CL exposes the /svc handle to the dart runtime.
This allows us to connect directly to the incoming
services instead of getting the from the environment proxy.

Doing this will allow us to, eventually, remove the
environment proxy from the startup context which
will require components to access services from the
/svc directory instead of through the environment.

We will also be able to bring the StartupContext class
more inline with the c++ version.

TEST=fuchsia_services_package_integration_tests

Change-Id: I22e0827359f3c931911f7ed13aaac6e78146c660