[media/mediaplayer] move Metadata and TimelineFunction, phase 2
This CL is the second of 5 moving fuchsia.mediaplayer.Metadata and
fuchsia.mediaplayer.TimelineFunction to fuchsia.media. Both of these
structs are media-generic, and fuchsia.mediasession needs to
reference them. These definitions are in garnet, hence the multistep
transition.

In this CL, we stop referencing the fields that use the
fuchsia.mediaplayer versions of the Metadata and TimelineFunction
structs so those versions can be removed in phase 3.

TEST: no behavior change
Change-Id: Iaf81cbe06b79ff58ee693d64ce4d9dcaca340775