[media/mediaplayer] move Metadata and TimelineFunction, phase 4
This CL is the fourth of 5 moving fuchsia.mediaplayer.Metadata and
fuchsia.mediaplayer.TimelineFunction to fuchsia.media. Both of these
structs are media-generic, and fuchsia.mediasession needs to
reference them. These definitions are referenced in topaz, hence
the multistep transition.

In this CL, we switch from using the alternate fields to the original
fields, which now have new fuchsia.media types. That will allow us
to remove the alternate fields in phase 5.

TEST: no behavior change
Change-Id: I285773240e4130b1019c75888b91ba4650243935