[flutter driver tests] Making hermetic; factoring away test driver
This change makes the flutter driver tests hermetic and factors away the
wrapping in the test driver module and test runner service. Locally at least,
this seems to remove the flakage. In a followup we'll want to add tests that
re-expose the legitimate problems discovered in DX-710.

DX-710 #done
DX-634 #done
IN-991 #done

Test: fx set x64 --available topaz/bundles/buildbot
      fx run-test driver_example_mod_tests
      fx run-test slider_mod_tests
      fx shell runtests
Change-Id: I6571b6415742804776eac7ccb2be42fabd7297ed