[inspect][dart] Instantiate the Inspect API in the example mod.
Add VmoWriter constructor that takes an int for VMO size: fixes runtime error in existing usage, which tries to pass an int to a ctor that takes a VmoHolder.
Add type annotation to Inspect constructor such that this can be statically checked by the Dart analyzer.

To be done in future CLs: hook up the Inspect API to actual data.

CF-602 #comment

Change-Id: I706e1112c9805407cf78363c88c96f2c9b7f62ea