[flutter_runner] Add f.d.m.Administrator access to flutter runners.
In 264852 I added support for the device_settings UI to call the
fuchsia.device.manager.Administrator service to reboot the device.

This works great on "release" builds for astro but not for the "debug"
builds which uses the "flutter_jit_runner.cmx" instead of the
"flutter_jit_product_runner.cmx".

This will add the ability to all the flutter_*_runner types, since this
is now part of the SDK then it makes sense that all types should be able
to access it.

Change-Id: I5e78a0ff20732422003dfc1f43af8b7f237f1e42