[sdk] Define a separate SDK for Dart additions to the Core SDK.
This allows these SDK contributions to be controlled via an API file.
Note that this will only be effective when the infrastructure is
updated to build the Core and "Dart" SDKs separately, as SDK targets are
not built by default anymore (and therefore verification is not by default
applied to the newly-created //topaz/public/sdk:fuchsia_dart).

Bug: DX-1200
Change-Id: I189dd721b0c27539082f575915c5e1cd04931681