[flutter test] Make flutter_test portable
This change relativizes the paths used in the generated
invocation scripts - and adds 'test spec' support to the target (meaning
that the build will register these targets as tests to pass onto
the infrastructure, running these in a shard on separate bot). Moreover,
this change stops supplying the test source directory as this is not
portable.

Bug: IN-823
Change-Id: Ib1a30d1a407a72cbc94c2550f1109ece8fe91fea