[fidl][dart] Test FIDL dangerous identifiers
This implements the FIDL dangerous identifiers test for Dart and fixes
some problems that the test found:
 - List, Map, override and toString should be escaped
 - callback names should use the transformed name
 - fields and methods on Enum types are prefixed with $ so as to not
   conflict with enum member names.

The reserved words list is now contextual so that a word that is
reserved as a global declaration (eg: List) would be allowed as a
protocol method name.

This test has allowed us to see that many identifiers that were formerly
mangled (eg: external) don't need to be mangled any more.

There's downstream code that uses EnumType.values and UnionType.tag so
to allow a soft transition EnumType.values is a synonym for
EnumType.$values, UnionType.tag is a synonym for UnionType.$tag, for
now. See FIDL-587.

FIDL-259
FIDL-107 #done

Change-Id: I316a433081b75ae04d6dd4dbcef94b07f4a5b137