[flutter_runner] Increase Skia graphics resource cache bytes to 16mb
This value started at 512mb, and was reduced to 8mb in efforts to fight
OOM issues that were happening at the time.  The value was not tuned
w.r.t. FPS.  For the image scrolling workload, just bumping it up to
16mb achieves significant "GPURasterizer::Draw" duration improvements.
The full stats from the traces I took are listed below.  Most
importantly, the reduction is large enough that the majority of our
"vsync callback" + "GPURasterizer::Draw" durations are now below the
16.67ms budget.

“GPURasterizer::Draw” durations:
 Before (8mb) -> After (16mb)
  avg 12,499,036 ns -> 3,258,264 ns
  max 32,923,333 ns -> 11,988,792 ns
  min 1,083,625 ns -> 1,203,375 ns
  std 3,663,162 ns -> 1,339,275 ns
  pct_050 11,565,301 ns -> 2,631,807 ns
  pct_090 16,556,576 ns -> 5,488,671 ns

See FL-153 for the traces.

It is possible that for different workloads, a larger cache size will be
required (in order to achieve 60 FPS), however for now let's try to
increase it by only the necessary amount.

FL-153 #comment

Change-Id: Iff56794e97aae964815f216c826e480c973d62b1