[sdk] Remove "topaz" aggregate SDK.
This SDK is simply the union of the "core" and "fuchsia_dart" SDK.
These two pieces are now merged together by our infra bots.
Later, DX-1438 will allow us to perform that merge step in the build
again in a way that does not interfere with API verification and matches
how merging will occur with prebuilt SDK add-ons.

Bug: DX-1238
Change-Id: Ic34a15a587572f9b61d87b3956209e3030132c22