[flutter] Fixes for the Flutter roll
This change also includes some additional checks in
dart:zircon a test that I added to help investigate
some tests that were transiently failing on the GI
bot. I'd like to retain them here, and add more such
checks in a subsequent topaz CL that isn't part of
the roll.

Change-Id: I566798b2d75b3ce46ada8f37af4a96ff9bbc9bda