[dart] Find unstripped files even for non-default variant
The profiler symbols action hardcodes filenames that assume the default
variant, but this breaks when non-default variant is being used, for
example ASan. Computing the correct filename from within GN is difficult
because we would need to execute the action from withing the variant
toolchain but there is no way to do that in current build. So instead
of hardcoding the filename of the unstripped file, locate the unstripped
file in the script that produces profiler symbols.

Change-Id: I8888eda517ede941a5aaf2e99804594d403c89c0