[flutter_runner] Improve frame scheduling
FL-233 #comment

This allows the paint tasks for the next frame to execute in parallel
with presentation of last frame but still provides back-pressure to
prevent us from queuing up even more work.

Vsync would be disabled whenever a presentation callback was pending
prior to this change. That had the outcome of causing us to almost
always miss one vsync interval. By not turning off vsync until
another Present call is pending we avoid this problem.

Test: fx shell run fuchsia-pkg://fuchsia.com/basemgr#meta/basemgr.cmx --base_shell=fuchsia-pkg://fuchsia.com/spinning_cube#meta/spinning_cube.cmx
Test: topaz input latency benchmarks
Test: end-2-end tests
Change-Id: I46440052cd4f98cb0992ec5027584be80f4fb9d3