[cleanup][sledge][setui] Fix Dart lints
Improvements in the latest Dart analyzer/linters discovered several
cases where we're missing return statements.

Sledge.runInTransaction() is expected to return a Future. For
consistency with surrounding code, we mark the block async.

Mockito's thenAnswer() is expected to return the result of the
invocation of the when() call; in this case the State that is the result
of the assignment/setter.

Change-Id: I6575a582dbfc269e1496a8ee0ce20900ad76d343