[Settings] Gate when the update channel source can be changed.
It was previously possible to select another update source while
changing the source was already in progress. This can lead
to multiple channels being active and/or an inconsistent state
shown on screen.

This changelist addresses this by preventing the channel
selection screen from showing when an update is in progress.
The button text is also changed to reflect that an update is
currently happening.

Fixes: SU-314
Test: fx run-host-tests device_settings_tests
Change-Id: If76d05d426508c944c5928daf7540f41cd458b89