[flutter_screencap_test] Move to flaky builders
It's quite flaky. This removes it from CQ.
The flaky builders run in CI and can be run as optional tryjobs, so
it should be possible to determine if the test has been fixed by
looking at the results from these builders.

Note this is the first usage of the flaky builders, set up as part of
IN-1130. If this works as expected, I'll document this process.

Bug: FLK-352
Change-Id: Ieec82e28ce88b06119db500e66edb830ec3a323a