[inspect][dart] Fixes heap grow bug
CF-897 #comment

The little big allocator heap grow was wrong in that it always
initialized the heap starting from the smalles indexes (1), even
if it was asked to double the heap in size.

This change fixes the issue, adds a regression test and adds a few
interesting assertions at critical points in the code, to increase
chance that heap corruption gets detected early.

Change-Id: Id5cfd5019c68e9374d3679529967173cd82adce6