[flutter_runner] Default observatory host
In https://github.com/flutter/engine/pull/9485, support was added for
specifying an observatory_host setting (along with the existing
Settings::observatory_port) and setting the default host was moved from
being hardcoded within the engine to being passed in via settings. This
value was initialised in the switches for runners in the engine repo.
This applies the same defaulting in flutter_runner.

This defaults the value to 127.0.0.1 which matches how it was set
implicitly in the past. Long-term this should be passed in as a parameter
via e.g. Application::Create() to allow for defaults other than
127.0.0.1.

Change-Id: I73ec3044d5590d88212d2f10a280fa52bc8d9cda