[crash] remove fuchsia.crash.Analyzer from list of services
I am pretty sure the service was added because of the warning message when a
component tries to connect to a service it doesn't explicitly list. In that
specific case, the issue was that when a component threw an unhandled Dart
exception, we were using the services of that component instead of the
runner's. This has been fixed in https://fuchsia-review.googlesource.com/c/topaz/+/242254!

DX-1820 #comment

Change-Id: I2f20f38d72a2d2110515749c5efac29b56643d2b