[fuchsia_inspect] Make integration test less brittle.
This test would fail when another component (archivist) came along and
read Inspect data. In general, diagnostics components will be able to hook
in and access arbitrary data in components' out directories, so this
test needed to be rewritten to no longer assume it would be the only
accessor of the component under test.

Change-Id: If1fab910811b86f9962e08e3171ea5915323d14c