[story shell] Add layout logic for a 'workspace' archetype
Add a 'workspace' archetype, with the following roles:
primary:    the main work element
auxiliary:  a helper surface (optional)
header:     the optional header surface (optional)
footer:     the optional footer surface (optional)

Auxiliaries can also specify a preferred visual hierarchy with respect
to the primary: either 'child' - the default, coming after the parent,
or 'parent' - preceeding the primary.

Surfaces are matched to the layout according to
a) declaring in their metadata that they are participating in the
   archetype and
b) the archetype_role they claim to be playing

Surfaces in specific roles can specify a preferred behavior for
multiplexing: e.g. three surfaces want to be 'auxiliary left', and the
most focused surface's expression wins (most recently added to story).
The behaviors are:
single: the surfaces are returned as a stack
toggle: the surfaces are returned as a toggleable, and the chrome to
        achieve that is left to the presenter
copresent: the surfaces are laid out using the copresentation algorithm
            _iff_ they declare a copresentation relation between
            themselves and the parent (e.g. the primary). This is
            subtle.

Add archetype_layout_test, for testing the different archetype
permutations

Test: fx run-host-tests composition_delegate_tests
Change-Id: If7b47a7b68db26deff26139e81e03533fc4dc24a