[doc_checker] adding check for full URLs
Links to documents in //docs should be relative to the root of the
fuchsia project. They should not assume any server address.

This change enforces this rule, and allows for exceptions in the case
of a list of files such as navbar.md which is a gtiles only file, and
pointing to the root of the fuchsia source tree.

Change-Id: Iba4f3e5037baf4b370d76421b0f7e1e23ae2dd5d