[fidl][dart] Compatibility test for writing unions as xunions
This introduces a second compatibility test server.
It is the same Dart program but launched with the argument
"write_xunions". When the test server program is launched with that
argument it sets the defaultEnableWriteXUnionBytesForUnion flag. This
server isn't yet introduced into the test matrix.

This CL also reduces the amount of integration other languages'
compatibility tests to simplify a clean-up I'm doing in fuchsia.git.

Change-Id: I326698bbd11d582afaeab5ab82b43392aa521127