[fidl][dart] Final fixes to make Dart compat green
(This incorporates changes which were started in
I51ec696e3a3e6fb4a954c9c7d1d39e7eae982452.)

We change the `Encoder` to take whether it should operate in
unions-as-xunions mode explicitely. The codegen is modified to use
the default value, and pass that explicitely.

The header is properly writtent with the flag set when writing
unions-as-xunions.

The bindings are changed to read the header bits,
indicating whether to use the unions-as-xunions on decode path.

Methods now have their old and v1 request/response sizes, which is then
used appropriately in the generated bindings code to allocate, position,
and claim buffer reads/writes.

Note: we change the compatibility test in the stacked CL since it also
relies on depedent CLs in fuchsia.git which have not yet globally
integrated.

Test: fx run-host-tests fidlgen_dart_backend_test
Change-Id: Ia42ea61e826d02c0babf08853c4afa0e91efff65