[testing] Disable non-component tests
The infrastructure will be looking to enable the use of
run-test-component for the main builders instead of runtests. There are
a number of pressing issues that this addresses (including a _drastic_
speed-up of testing time) - and accordingly it is more utilitarian to
disable the tests that would prevent this within topaz.

Bug: 8845
Bug: 8846
Bug: 10427

Change-Id: Idb4c383b6e3a98d9c740282b2770e5deef0f6167