[fidl] Read both hashed and explicit ordinals
This updates the dart bindings to accept the two sets of
ordinals specified in the "hashed_ordinal" and "explicit_ordinal"
fields from the JSON IR  when decoding xunions.

After decoding, the ordinal found "on the wire" is discarded
and the actual value stored in the xunion is the one that is
specified in the "ordinal" field in the IR (which is currently
the hashed ordinal for most xunions and the explicit ordinal
for all static unions that are now xunions). This internal
ordinal field is also used when encoding the xunion.

Once all bindings support reading both sets of unions, the
"ordinal" field for all unions/xunions will be set to the
explicit ordinal.

Test: fx run-test fidl_bindings_test
Change-Id: Ib3b59b557ea29c39489893399cb092e2e7b72e13