[fidl] Remove runtime flags
Bug: 42311

This removes the encodeUnionAsXUnionBytes and
decodeUnionFromXUnionBytes runtime flags from the Encoder and
Decoder respectively and all the code that relied on these
flags. Since we are already writing v1 for everything, this
CL has no behavior change.
This CL also cleans up the topaz compatibility test naming by
removing the _write_xunion suffix.

Test: fx run-test fidl_bindings_test
Test: fx run-test fidl_compatibility_test_topaz
Change-Id: I3bc7615bd225e3f9ec245e1a926e298d67b60ebe