[copy][ninja] Move copy_sky_engine_sources.py to Python3
Python2 doesn't support nanosecond timestamps, but ninja does for
checking dirtiness of builds. Because python2 is being used to copy the
sources from the dart dir to the output dir in Fuchsia builds, doing a
second build immediately after the first one results in the build still
being dirty and the steps repeated.

In order to fix this, simply update this script to Python 3.x which
handles nanoseconds in mtimes.

Bug: 36581
Change-Id: I77cad2ec6b5b0f1d290bc1aeecf259462884338a