[fidlgen_dart] Remove dep on cmdline package
Removing depedency on the cmdline package, which is defunct part of the
"fidlgen lib" world.

The cmdline package was used to define "standard flags" for all fidlgen.
However, none of the backends need the same input since they all differ
depending on what is needed or supported for the various target
languages.

Instead, defining special purpose flags appropriate for Dart bindings
generation.

Also relying on the types package to read in the JSON IR.

Improvement to the usage in line with what is done in the Rust backend.

Test: build integration, codegen tests
Change-Id: I20b2d5c60a398183fea69399deab4301079406fb